backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.1
python-dotenv==1.0.0
requests==2.31.0
pymongo==4.6.1
